refactor: extract shared auth base class for page components
