+++ WSIST/WSIST.Web/Components/Pages/AuthenticatedComponentBase.cs
@@ -1,0 +1,39 @@
+using System.Security.Claims;
+using Microsoft.AspNetCore.Components;
+using Microsoft.AspNetCore.Components.Authorization;
+using WSIST.Engine;
+
+namespace WSIST.Web.Components.Pages;
+
+public abstract class AuthenticatedComponentBase(
+    TestManagement management,
+    AuthenticationStateProvider authStateProvider,
+    NavigationManager navigation) : ComponentBase
+{
+    protected int CurrentUserId { get; private set; }
+
+    protected override async Task OnInitializedAsync()
+    {
+        var authState = await authStateProvider.GetAuthenticationStateAsync();
+        var user = authState.User;
+
+        if (user?.Identity?.IsAuthenticated != true)
+        {
+            navigation.NavigateTo("/login-page", forceLoad: true);
+            return;
+        }
+
+        var email = user.FindFirst(ClaimTypes.Email)?.Value;
+        var name = user.FindFirst(ClaimTypes.Name)?.Value ?? "Unknown";
+        var googleId = user.FindFirst(ClaimTypes.NameIdentifier)?.Value ?? "";
+
+        if (email is null) return;
+
+        var dbUser = management.GetOrCreateUser(email, name, googleId);
+        CurrentUserId = dbUser.Id;
+
+        await OnAuthenticatedAsync();
+    }
+
+    protected virtual Task OnAuthenticatedAsync() => Task.CompletedTask;
+}
