+++ WSIST/WSIST.Web/Components/Pages/Home.razor
@@ -1,6 +1,7 @@
 ﻿@page "/"
 @rendermode @(new InteractiveServerRenderMode(prerender: false))
 @using WSIST.Engine
+@inherits AuthenticatedComponentBase
 <PageTitle>WSIST - Home</PageTitle>
 
 <div class="page">
