+++ WSIST/WSIST.Web/Components/Pages/Home.razor.cs
@@ -6,34 +6,18 @@ using WSIST.Engine;
 namespace WSIST.Web.Components.Pages;
 
 public partial class Home(TestManagement management, AuthenticationStateProvider authStateProvider, NavigationManager navigation)
+    : AuthenticatedComponentBase(management, authStateProvider, navigation)
 {
     private List<Test> tests = [];
     private Test? temporaryTest;
     private Modes Mode { get; set; }
-    private int currentUserId;
 
-    protected override async Task OnInitializedAsync()
+    protected override Task OnAuthenticatedAsync()
     {
-        var authState = await authStateProvider.GetAuthenticationStateAsync();
-        var user = authState.User;
-
-        if (user?.Identity?.IsAuthenticated != true)
-        {
-            navigation.NavigateTo("/login-page", forceLoad: true);
-            return;
-        }
-
-        var email = user.FindFirst(ClaimTypes.Email)?.Value;
-        var name = user.FindFirst(ClaimTypes.Name)?.Value ?? "Unknown";
-        var googleId = user.FindFirst(ClaimTypes.NameIdentifier)?.Value ?? "";
-
-        if (email is null) return;
-
-        var dbUser = management.GetOrCreateUser(email, name, googleId);
-        currentUserId = dbUser.Id;
-        tests = management.LoadAllTests(currentUserId)
+        tests = management.LoadAllTests(CurrentUserId)
             .Where(t => t.DueDate >= DateOnly.FromDateTime(DateTime.Today))
             .ToList();
+        return Task.CompletedTask;
     }
 
     public enum Modes
@@ -86,7 +70,7 @@ public partial class Home(TestManagement management, AuthenticationStateProvider
                     temporaryTest.Volume,
                     temporaryTest.Understanding,
                     temporaryTest.Grade,
-                    currentUserId
+                    CurrentUserId
                 );
                 break;
             }
@@ -116,7 +100,7 @@ public partial class Home(TestManagement management, AuthenticationStateProvider
 
     private void Refresh()
     {
-        tests = management.LoadAllTests(currentUserId)
+        tests = management.LoadAllTests(CurrentUserId)
             .Where(t => t.DueDate >= DateOnly.FromDateTime(DateTime.Today))
             .ToList();
         StateHasChanged();
