+++ WSIST/WSIST.Web/Components/Pages/Study.razor
@@ -1,5 +1,6 @@
 ﻿@page "/study"
 @rendermode @(new InteractiveServerRenderMode(prerender: false))
+@inherits AuthenticatedComponentBase
 <PageTitle>WSIST - Study</PageTitle>
 
 <div class="page">
